requirements.txt
Flask==2.2.2
numpy==1.23.0
pandas==1.4.3
tensorflow==2.10.0
tensorflow-estimator==2.10.0
Werkzeug==2.2.2
Keras-Preprocessing==1.1.2
gunicorn==20.1.0
regex